simplex-alerter/pyproject.toml
[project]
name = "simplex-alerter"
version = "1.4.0"
description = "webhook based simpleX alerter"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "observlib",
    "opentelemetry-api>=1.33.1",
    "opentelemetry-instrumentation-asyncio>=0.53b1",
    "opentelemetry-instrumentation-fastapi>=0.54b1",
    "opentelemetry-instrumentation-threading>=0.53b1",
    "opentelemetry-sdk>=1.33.1",
    "prometheus-client>=0.22.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
    "opentelemetry-instrumentation-asgi>=0.54b1",
    "jinja2>=3.1.6",
    "websockets>=15.0.1",
    "pexpect>=4.9.0",
]

[project.scripts]
simplex-alerter = "simplex_alerter:__main__.run"

[tool.uv.sources]
observlib = { git = "https://github.com/ForgottenBeast/observlib.git", subdirectory = "observlib" }
